Refactor to make error handler ordering explicit
Rather than making order driven by definition order, expose the
declared hook type (was `_RegisteredHook`), and convert the decorated
functions into the declared hook objects. Then, ordering can be done
by any sorting function -- currently, a strict ordered listing with no
implicit ordering semantics -- and we can explicitly register the
results.

This allows us to break the giant hook module down into its
constituent parts which can be further refined and reorganized over
time.
