--- src/globus_cli/exception_handling/hooks.py
@@ -1,410 +1,0 @@
-from __future__ import annotations
-
-import json
-import textwrap
-import typing as t
-
-import click
-import globus_sdk
-import globus_sdk.gare
-
-from globus_cli.endpointish import WrongEntityTypeError
-from globus_cli.login_manager import MissingLoginError
-from globus_cli.termio import PrintableErrorField, outformat_is_json, write_error_info
-from globus_cli.types import JsonValue
-from globus_cli.utils import CLIAuthRequirementsError
-
-from .messages import (
-    DEFAULT_CONSENT_REAUTH_MESSAGE,
-    DEFAULT_SESSION_REAUTH_MESSAGE,
-    emit_consent_required_message,
-    emit_message_for_gare,
-    emit_session_update_message,
-    emit_unauthorized_message,
-)
-from .registry import error_handler, sdk_error_handler
-
-
-def _pretty_json(data: JsonValue, compact: bool = False) -> str:
-    if compact:
-        return json.dumps(data, separators=(",", ":"), sort_keys=True)
-    return json.dumps(data, indent=2, separators=(",", ": "), sort_keys=True)
-
-
-_JSONPATH_SPECIAL_CHARS = "[]'\"\\."
-_JSONPATH_ESCAPE_MAP = str.maketrans(
-    {
-        "'": "\\'",
-        "\\": "\\\\",
-    }
-)
-
-
-def _jsonpath_from_pydantic_loc(loc: list[str | int]) -> str:
-    """
-    Given a 'loc' from pydantic error data, convert it into a JSON Path expression.
-
-    Takes the following steps:
-    - turns integers into integer indices
-    - turns most strings into dotted access
-    - turns strings containing special characters into single-quoted bracket access
-      with ' and \\ escaped
-    """
-    path = "$"
-    for part in loc:
-        if isinstance(part, int):
-            path += f"[{part}]"
-        else:
-            if any(c in part for c in _JSONPATH_SPECIAL_CHARS):
-                part = f"'{part.translate(_JSONPATH_ESCAPE_MAP)}'"
-                path += f"[{part}]"
-            else:
-                path += f".{part}"
-    return path
-
-
-@sdk_error_handler(
-    error_class="GlobusAPIError", condition=lambda err: err.raw_json is None
-)
-def null_data_error_handler(exception: globus_sdk.GlobusAPIError) -> None:
-    write_error_info(
-        "GlobusAPINullDataError",
-        [PrintableErrorField("error_type", exception.__class__.__name__)],
-    )
-
-
-@sdk_error_handler(
-    error_class="GlobusAPIError", condition=lambda err: outformat_is_json()
-)
-def json_error_handler(exception: globus_sdk.GlobusAPIError) -> None:
-    msg = json.dumps(exception.raw_json, indent=2, separators=(",", ": "))
-    click.secho(msg, fg="yellow", err=True)
-
-
-@error_handler(error_class=CLIAuthRequirementsError, exit_status=4)
-def handle_internal_auth_requirements(
-    exception: CLIAuthRequirementsError,
-) -> int | None:
-    gare = exception.gare
-    if not gare:
-        msg = "Fatal Error: Unsupported internal auth requirements error!"
-        click.secho(msg, bold=True, fg="red")
-        return 255
-
-    emit_message_for_gare(gare, exception.message)
-
-    if exception.epilog:
-        click.echo("\n* * *\n")
-        click.echo(exception.epilog)
-
-    return None
-
-
-@sdk_error_handler(
-    error_class="FlowsAPIError",
-    condition=lambda err: globus_sdk.gare.is_gare(err.raw_json or {}),
-    exit_status=4,
-)
-def handle_flows_gare(exception: globus_sdk.FlowsAPIError) -> int | None:
-    gare = globus_sdk.gare.to_gare(exception.raw_json or {})
-    if not gare:
-        raise ValueError("Expected a GARE, but got None")
-
-    emit_message_for_gare(gare)
-    return None
-
-
-@sdk_error_handler(
-    condition=lambda err: bool(err.info.authorization_parameters), exit_status=4
-)
-def session_hook(exception: globus_sdk.GlobusAPIError) -> None:
-    """
-    Expects an exception with a valid authorization_paramaters info field.
-    """
-    message = exception.info.authorization_parameters.session_message
-    if message:
-        message = f"{DEFAULT_SESSION_REAUTH_MESSAGE}\nmessage: {message}"
-    else:
-        message = DEFAULT_SESSION_REAUTH_MESSAGE
-
-    emit_session_update_message(
-        identities=exception.info.authorization_parameters.session_required_identities,
-        domains=exception.info.authorization_parameters.session_required_single_domain,
-        policies=exception.info.authorization_parameters.session_required_policies,
-        message=message,
-    )
-    return None
-
-
-@sdk_error_handler(condition=lambda err: bool(err.info.consent_required), exit_status=4)
-def consent_required_hook(exception: globus_sdk.GlobusAPIError) -> int | None:
-    """
-    Expects an exception with a required_scopes field in its raw_json.
-    """
-    if not exception.info.consent_required.required_scopes:
-        msg = "Fatal Error: ConsentRequired but no required_scopes!"
-        click.secho(msg, bold=True, fg="red")
-        return 255
-
-    # specialized message for data_access errors
-    # otherwise, use more generic phrasing
-    if exception.message == "Missing required data_access consent":
-        message = (
-            "The collection you are trying to access data on requires you to "
-            "grant consent for the Globus CLI to access it."
-        )
-    else:
-        message = f"{DEFAULT_CONSENT_REAUTH_MESSAGE}\nmessage: {exception.message}"
-
-    emit_consent_required_message(
-        required_scopes=exception.info.consent_required.required_scopes, message=message
-    )
-    return None
-
-
-@sdk_error_handler(
-    condition=lambda err: (
-        (
-            isinstance(err, globus_sdk.TransferAPIError)
-            and err.code == "ClientError.AuthenticationFailed"
-        )
-        or (isinstance(err, globus_sdk.AuthAPIError) and err.code == "UNAUTHORIZED")
-    )
-)
-def authentication_hook(
-    exception: globus_sdk.TransferAPIError | globus_sdk.AuthAPIError,
-) -> None:
-    emit_unauthorized_message()
-
-
-@sdk_error_handler(error_class="TransferAPIError")
-def transferapi_hook(exception: globus_sdk.TransferAPIError) -> None:
-    write_error_info(
-        "Transfer API Error",
-        [
-            PrintableErrorField("HTTP status", exception.http_status),
-            PrintableErrorField("request_id", exception.request_id),
-            PrintableErrorField("code", exception.code),
-            PrintableErrorField("message", exception.message, multiline=True),
-        ],
-    )
-
-
-@sdk_error_handler(
-    error_class="SearchAPIError",
-    condition=lambda err: err.code == "BadRequest.ValidationError",
-)
-def searchapi_validationerror_hook(exception: globus_sdk.SearchAPIError) -> None:
-    fields = [
-        PrintableErrorField("HTTP status", exception.http_status),
-        PrintableErrorField("request_id", exception.request_id),
-        PrintableErrorField("code", exception.code),
-        PrintableErrorField("message", exception.message, multiline=True),
-    ]
-    error_data: dict[str, JsonValue] | None = exception.error_data
-    if error_data is not None:
-        messages = error_data.get("messages")
-        if isinstance(messages, dict) and len(messages) == 1:
-            error_location, details = next(iter(messages.items()))
-            fields += [
-                PrintableErrorField("location", error_location),
-                PrintableErrorField("details", _pretty_json(details), multiline=True),
-            ]
-        elif messages is not None:
-            fields += [
-                PrintableErrorField("details", _pretty_json(messages), multiline=True)
-            ]
-
-    write_error_info("Search API Error", fields)
-
-
-@sdk_error_handler(error_class="SearchAPIError")
-def searchapi_hook(exception: globus_sdk.SearchAPIError) -> None:
-    fields = [
-        PrintableErrorField("HTTP status", exception.http_status),
-        PrintableErrorField("request_id", exception.request_id),
-        PrintableErrorField("code", exception.code),
-        PrintableErrorField("message", exception.message, multiline=True),
-    ]
-    error_data: dict[str, JsonValue] | None = exception.error_data
-    if error_data is not None:
-        fields += [
-            PrintableErrorField("error_data", _pretty_json(error_data, compact=True))
-        ]
-
-    write_error_info("Search API Error", fields)
-
-
-@sdk_error_handler(
-    error_class="FlowsAPIError",
-    condition=lambda err: err.code == "UNPROCESSABLE_ENTITY",
-)
-def flows_validation_error_hook(exception: globus_sdk.FlowsAPIError) -> None:
-    message_string: str = exception.raw_json["error"]["message"]  # type: ignore[index]
-    details: str | list[dict[str, t.Any]] = exception.raw_json["error"]["detail"]  # type: ignore[index] # noqa: E501
-    message_fields = [PrintableErrorField("message", message_string)]
-
-    # conditionally do this work if there are multiple details
-    if isinstance(details, list) and len(details) > 1:
-        num_errors = len(details)
-        # try to extract 'loc' and 'msg' from the details, but only
-        # update 'message_fields' if the data are present
-        try:
-            messages = [
-                f"{_jsonpath_from_pydantic_loc(data['loc'])}: {data['msg']}"
-                for data in details
-            ]
-        except KeyError:
-            pass
-        else:
-            message_fields = [
-                PrintableErrorField(
-                    "message", f"{num_errors} validation errors", multiline=True
-                ),
-                PrintableErrorField(
-                    "errors",
-                    "\n".join(messages),
-                    multiline=True,
-                ),
-            ]
-
-    write_error_info(
-        "Flows API Error",
-        [
-            PrintableErrorField("HTTP status", exception.http_status),
-            PrintableErrorField("code", exception.code),
-            *message_fields,
-        ],
-    )
-
-
-@sdk_error_handler(error_class="FlowsAPIError")
-def flows_error_hook(exception: globus_sdk.FlowsAPIError) -> None:
-    assert exception.raw_json is not None  # Influence mypy's knowledge of `raw_json`.
-    details: list[dict[str, t.Any]] | str = exception.raw_json["error"]["detail"]
-    detail_fields: list[PrintableErrorField] = []
-
-    # if the detail is a string, return that as a single field
-    if isinstance(details, str):
-        if len(details) > 120:
-            details = textwrap.fill(details, width=80)
-        detail_fields = [PrintableErrorField("detail", details, multiline=True)]
-    # if it's a list of objects, wrap them into a multiline detail field
-    elif isinstance(details, list):
-        num_errors = len(details)
-        if all((isinstance(d, dict) and "loc" in d and "msg" in d) for d in details):
-            detail_strings = [
-                (
-                    ((data["type"] + " ") if "type" in data else "")
-                    + f"{_jsonpath_from_pydantic_loc(data['loc'])}: {data['msg']}"
-                )
-                for data in details
-            ]
-            if num_errors == 1:
-                detail_fields = [PrintableErrorField("detail", detail_strings[0])]
-            else:
-                detail_fields = [
-                    PrintableErrorField("detail", f"{num_errors} errors"),
-                    PrintableErrorField(
-                        "errors",
-                        "\n".join(detail_strings),
-                        multiline=True,
-                    ),
-                ]
-        else:
-            detail_fields = [
-                PrintableErrorField(
-                    "detail",
-                    "\n".join(_pretty_json(detail, compact=True) for detail in details),
-                    multiline=True,
-                )
-            ]
-
-    fields = [
-        PrintableErrorField("HTTP status", exception.http_status),
-        PrintableErrorField("code", exception.code),
-    ]
-    if "message" in exception.raw_json["error"]:
-        fields.append(
-            PrintableErrorField("message", exception.raw_json["error"]["message"])
-        )
-    fields.extend(detail_fields)
-
-    write_error_info("Flows API Error", fields)
-
-
-@sdk_error_handler(
-    error_class="AuthAPIError",
-    condition=lambda err: err.message == "invalid_grant",
-)
-def invalidrefresh_hook(exception: globus_sdk.AuthAPIError) -> None:
-    emit_unauthorized_message()
-
-
-@sdk_error_handler(error_class="AuthAPIError")
-def authapi_hook(exception: globus_sdk.AuthAPIError) -> None:
-    write_error_info(
-        "Auth API Error",
-        [
-            PrintableErrorField("HTTP status", exception.http_status),
-            PrintableErrorField("code", exception.code),
-            PrintableErrorField("message", exception.message, multiline=True),
-        ],
-    )
-
-
-@sdk_error_handler()  # catch-all
-def globusapi_hook(exception: globus_sdk.GlobusAPIError) -> None:
-    write_error_info(
-        "Globus API Error",
-        [
-            PrintableErrorField("HTTP status", exception.http_status),
-            PrintableErrorField("code", exception.code),
-            PrintableErrorField("message", exception.message, multiline=True),
-        ],
-    )
-
-
-@sdk_error_handler(error_class="GlobusError")
-def globus_error_hook(exception: globus_sdk.GlobusError) -> None:
-    write_error_info(
-        "Globus Error",
-        [
-            PrintableErrorField("error_type", exception.__class__.__name__),
-            PrintableErrorField("message", str(exception), multiline=True),
-        ],
-    )
-
-
-@error_handler(error_class=WrongEntityTypeError, exit_status=3)
-def wrong_endpoint_type_error_hook(exception: WrongEntityTypeError) -> None:
-    msg = exception.expected_message + "\n" + exception.actual_message + "\n\n"
-    click.secho(msg, fg="yellow", err=True)
-
-    should_use = exception.should_use_command()
-    if should_use:
-        click.echo(
-            "Please run the following command instead:\n\n"
-            f"    {should_use} {exception.endpoint_id}\n",
-            err=True,
-        )
-    else:
-        msg = "This operation is not supported on objects of this type."
-        click.secho(msg, fg="red", bold=True, err=True)
-
-
-@error_handler(error_class=MissingLoginError, exit_status=4)
-def missing_login_error_hook(exception: MissingLoginError) -> None:
-    click.echo(
-        click.style("MissingLoginError: ", fg="yellow") + exception.message,
-        err=True,
-    )
-
-
-def register_all_hooks() -> None:
-    """
-    This is a stub method which does nothing.
-
-    Importing and running it serves to ensure that the various hooks were imported,
-    however. It therefore "looks imperative" and ensures that the hooks are loaded.
-    """
